+++ src/relstorage/adapters/sqlite/tests/test_dialect.py
@@ -1,0 +1,45 @@
+# -*- coding: utf-8 -*-
+##############################################################################
+#
+# Copyright (c) 2019 Zope Foundation and Contributors.
+# All Rights Reserved.
+#
+# This software is subject to the provisions of the Zope Public License,
+# Version 2.1 (ZPL).  A copy of the ZPL should accompany this distribution.
+# THIS SOFTWARE IS PROVIDED "AS IS" AND ANY AND ALL EXPRESS OR IMPLIED
+# WARRANTIES ARE DISCLAIMED, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF TITLE, MERCHANTABILITY, AGAINST INFRINGEMENT, AND FITNESS
+# FOR A PARTICULAR PURPOSE.
+#
+##############################################################################
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from relstorage.adapters.sql.tests import test_sql
+
+from .. import dialect
+
+class TestSQLiteUpsertDialect(test_sql.TestUpsert):
+    keep_history = False
+    dialect = dialect.Sqlite3Dialect()
+
+    def test_default(self):
+        self.assertEqual(
+            dialect.SQ3_SUPPORTS_UPSERT,
+            dialect.SqliteCompiler is dialect._Sqlite3UpsertCompiler)
+
+    insert_or_replace = test_sql.TestUpsert.insert_or_replace.replace(
+        '%s', '?'
+    )
+
+    insert_or_replace_subquery = test_sql.TestUpsert.insert_or_replace_subquery.replace(
+        '%s', '?'
+    )
+
+    upsert_unconstrained_subquery = test_sql.TestUpsert.upsert_unconstrained_subquery.replace(
+        '%s', '?'
+    ).replace(
+        ' ON ', ' WHERE true ON '
+    )